fix unrenderable bugs, restructure, add BatchTask feature, v0.1.7
