+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="ffmpeg_toolkit",
-    version="0.1.6",
+    version="0.1.7",
     author="Wenyang Tai",
     author_email="superyngo@gmail.com",
     description="A toolkit for working with FFmpeg",
